Jieun

+++ petmily/src/main/java/com/kh/petmily/entity/AccountDto.java
@@ -17,7 +17,7 @@ public class AccountDto {
 	private String account_sitter_id; //pet_sitter 테이블의 sitter_id
 	private int account_count; // 전월 결제 건 수 
 	private int account_total_pay; // 전월 토탈 금액
-	private int account_fees; // 전월 적용 수수료
+	private double account_fees; // 전월 적용 수수료
 	private String account_date; // 전월 정산 일자
